+++ Assets/OkapiKit/Scripts/Actions/ActionQuestGive.cs
@@ -1,0 +1,58 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using NaughtyAttributes;
+
+namespace OkapiKit
+{
+    [AddComponentMenu("Okapi/Action/Quest Give")]
+    public class ActionQuestGive : Action
+    {
+        [SerializeField]
+        private TargetQuestManager  questManager;
+        [SerializeField]
+        private Quest               quest;
+
+        public override string GetActionTitle() { return "Give Quest"; }
+
+        public override string GetRawDescription(string ident, GameObject gameObject)
+        {
+            string desc = GetPreconditionsString(gameObject);
+
+            var questName = quest?.displayName ?? "UNDEFINED";
+            var targetName = questManager.GetShortDescription(gameObject);
+            if (targetName != "") targetName = $"to {targetName}";
+            desc += $"Give quest \"{questName}\" {targetName}";
+
+            return desc;
+        }
+
+        protected override void CheckErrors()
+        {
+            base.CheckErrors();
+
+            if (quest == null)
+            {
+                _logs.Add(new LogEntry(LogEntry.Type.Warning, "Undefined target quest!", "You need to define the quest to give!"));
+            }
+            questManager.CheckErrors(_logs, "quest manager", gameObject);
+        }
+
+        public override void Execute()
+        {
+            if (!enableAction) return;
+            if (!EvaluatePreconditions()) return;
+
+            var qm = questManager.GetTarget(gameObject);
+            if (qm == null)
+            {
+                Debug.LogError("Can't find quest manager!");
+                return;
+            }
+            if (quest != null)
+            {
+                qm.AddQuest(quest);
+            }
+        }
+    }
+}